+++ src/Method/Traits/EmojisVariableTrait.php
@@ -13,6 +13,9 @@ trait EmojisVariableTrait
      * One or more emoji corresponding to the sticker.
      *
      * @var string
+     *
+     * @deprecated The Bot API moved this into InputSticker::emoji_list. Set
+     *             InputStickerType::$emojiList on the sticker instead.
      */
     public $emojis;
 }